Rewrite camera math

+++ src/main/java/legend/game/combat/environment/CameraQuadParamCallback.java
@@ -2,5 +2,5 @@ package legend.game.combat.environment;
 
 @FunctionalInterface
 public interface CameraQuadParamCallback {
-  void accept(int x, int y, int z, int scriptIndex);
+  void accept(float x, float y, float z, int scriptIndex);
 }
